table_buffer is now building
Took the task of building the table away from screen_buffer and gave it
to table_buffer. table_buffer is in charge of creating the table, while
screen_buffer is in charge of adding special effects for the screen.
Right now it doesn't do anything and might be a waste.

Might make it do something more important in the future.